int to float conversion

+++ options.h
@@ -5,7 +5,7 @@
 #ifndef OPTIONSPRICING_OPTIONS_H
 #define OPTIONSPRICING_OPTIONS_H
 
-int call_payoff (int strike_price, int stock_price) {
+float call_payoff (float strike_price, float stock_price) {
     /* call payoff is defined as: Ct = max(0, St - K), where
      * Ct denotes the payoff at time t, St the stock price
      * at maturity and K the strike price. */
